refactor(cli)!: rename dap_cli package to arca_flow_cli
Rename the CLI's Python package directory and distribution name as the
first step of Phase 1 of the arca-flow migration. The binary is still
named `dap` (renamed separately in the next commit) and user-facing
strings still mention DAP; only the import path and wheel name change
here.

BREAKING CHANGE: The distribution name is now `arca-flow-cli` and the
import path is `arca_flow_cli`. External consumers importing `dap_cli`
must update their imports.
