+++ cli/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["hatchling>=1.12.0"]
 build-backend = "hatchling.build"
 
 [project]
-name = "dap-cli"
+name = "arca-flow-cli"
 version = "0.1.0"
 description = "Developer CLI for the Data Archive Pipeline (DAP) - Orchestrator"
 requires-python = ">=3.12"
@@ -14,4 +14,4 @@ dependencies = [
 ]
 
 [project.scripts]
-dap = "dap_cli.app:main"
+dap = "arca_flow_cli.app:main"
